# specs/001-wiki-editor-backend/contracts/public-api.openapi.yaml
openapi: 3.1.0
info:
  title: Wiki Editor Backend Public API
  version: 0.2.0
  description: Public REST contracts exposed through the API Gateway.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /pages/{pageId}/publish:
    post:
      tags: [Pages]
      summary: Publish current draft
      operationId: publishPage
      parameters:
        - $ref: '#/components/parameters/RequestIdHeader'
        - $ref: '#/components/parameters/PageId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishPageRequest' }
      responses:
        '200':
          description: Publish accepted
          headers:
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PagePublishResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  headers:
    RequestId:
      description: Correlation and request tracking identifier
      schema:
        type: string
  parameters:
    RequestIdHeader:
      in: header
      name: X-Request-Id
      required: false
      schema:
        type: string
    PageId:
      in: path
      name: pageId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Forbidden:
      description: Authorization failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: State conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
        request_id: { type: string }
...
    PublishPageRequest:
      type: object
      required: [base_revision_no]
      properties:
        base_revision_no: { type: integer, format: int64 }
    PagePublishResponse:
      type: object
      properties:
        page_id: { type: string, format: uuid }
        published_revision_no: { type: integer, format: int64 }
        published_revision_id: { type: string, format: uuid }
        published_at: { type: string, format: date-time }
...